+++ packages/realm/src/platform/node/index.ts
@@ -18,5 +18,6 @@
 
 import "./fs";
 import "./device-info";
+import "./sync-proxy-config";
 
 export * from "../../index";
